msm: camera: Add support to VFE driver for 7x27a.
VFE driver interacts with the VFETASK in ADSP. It handles
the communication with this task. It sends the commands
to VFETASK and processes the messages received from
VFETASK and dispatches them to msm_camera. It also
handles the voting of required ebi1_vfe_clk rate based
on preview and snapshot requirements.

Change-Id: I592a8d1560a77498e0ed09fd6fd2ddbade73eea3
Signed-off-by: Suresh Vankadara <svankada@codeaurora.org>